@@ -1,52 +1,78 @@
 {
-  "name": "Fms",
-  "version": "1.0.0",
+  "name": "adsbold-webapp",
+  "version": "2.0.0",
   "description": "Facebook management suite",
   "scripts": {
-    "dev": "webpack -w --progress --config webpack.dev.js",
-    "build": "webpack --progress --config webpack.dev.js",
-    "build-prod": "NODE_ENV=production webpack --progress --config webpack.prod.js",
-    "deploy": "NODE_ENV=production webpack --progress --config webpack.prod.js && ./script/deploy",
     "prestart": "npm i",
-    "start": "node server.js"
+    "clean": "rimraf build",
+    "start": "node server.js",
+    "local": "npm run clean && webpack -w --progress --config config/webpack/webpack.local.js",
+    "local:alz": "npm run clean && webpack --progress --config config/webpack/webpack.local-alz.js",
+    "build-local": "NODE_ENV=local webpack --progress --config config/webpack/webpack.local.js",
+    "build-dev": "NODE_ENV=dev webpack --progress --config config/webpack/webpack.dev.js",
+    "build-staging": "NODE_ENV=staging webpack --progress --config config/webpack/webpack.staging.js",
+    "build-prod": "NODE_ENV=production webpack --progress --config config/webpack/webpack.prod.js"
   },
-  "author": "Fms team",
+  "author": "Adsbold team",
   "license": "MIT",
   "dependencies": {
+    "animate.css": "^3.5.2",
     "axios": "^0.9.1",
     "bootstrap": "^3.3.7",
+    "create-react-class": "^15.6.2",
+    "font-awesome": "^4.7.0",
     "form-data": "^2.3.1",
-    "jquery": "^2.2.1",
-    "muicss": "^0.9.27",
-    "react": "^0.14.7",
+    "jquery": "^2.2.4",
+    "jquery-flot": "^0.8.3",
+    "jquery-sparkline": "^2.4.0",
+    "metismenu": "^2.7.1",
+    "moment-mini": "^2.19.4",
+    "prop-types": "^15.6.0",
+    "react": "^16.2.0",
     "react-bootstrap": "^0.31.2",
+    "react-bootstrap-daterangepicker": "^3.4.1",
     "react-bs-notifier": "^4.3.2",
-    "react-dom": "^0.14.7",
-    "react-router": "^2.8.1",
+    "react-dom": "^16.2.0",
+    "react-loadable": "^5.3.1",
+    "react-redux": "^5.0.6",
     "react-router-dom": "^4.2.2",
+    "react-sparklines": "^1.7.0",
+    "redux": "^3.7.2",
+    "redux-thunk": "^2.2.0",
     "socket.io-client": "^2.0.3",
+    "sparkline": "^0.2.0",
     "store": "^2.0.12",
     "url-search-params-polyfill": "^2.0.1",
     "uuid": "^3.1.0"
   },
   "devDependencies": {
-    "webpack": "^3.7.1",
-    "sass-loader": "^3.1.2",
-    "json-loader": "^0.5.7",
+    "babel": "^6.23.0",
     "babel-core": "^6.5.1",
     "babel-loader": "^7.1.1",
-    "babel-preset-es2015": "^6.5.0",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-0": "^6.5.0",
-    "node-sass": "^3.4.2",
+    "core-js": "^2.5.3",
     "css-loader": "^0.23.1",
     "express": "^4.13.4",
     "extract-loader": "^1.0.1",
+    "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^0.11.2",
     "html-loader": "^0.5.1",
+    "html-webpack-plugin": "^2.30.1",
+    "json-loader": "^0.5.7",
+    "node-sass": "^4.7.2",
+    "react-hot-loader": "^3.1.3",
+    "rimraf": "^2.6.2",
+    "rxjs": "^5.5.2",
+    "sass-loader": "^3.1.2",
     "script-loader": "^0.6.1",
     "style-loader": "^0.13.0",
+    "suppress-chunks-webpack-plugin": "0.0.4",
+    "uglifyjs-webpack-plugin": "^1.1.1",
     "url-loader": "^0.5.9",
+    "webpack": "^3.7.1",
     "webpack-merge": "^4.1.0"
   }
 }
