@@ -1,113 +1,78 @@
-'use strict';
-
-import store from 'store';
 import axios from 'axios';
-import config from 'CONFIG';
-import DashboardAPI from 'DashboardApi';
+import {BASE_URL} from 'CONFIG';
+import tokenGetter from 'helpers/token-getter';
 
-const BASE_URL = config.BASE_URL;
-
-exports.get = (route, access_token) => {
-  let jwt = store.get('jwt');
-  let url = `${BASE_URL}${route}`;
-  let headers = {
-    'authorization': access_token || jwt
-  }
-
-  return axios.get(url, {headers})
-    .then(res => {
-      if (!res.data) {
-        throw new Error('Something went wrong');
-      } else {
-        if (res.data.err) {
-          throw new Error(res.data.msg);
-        } else {
-          return Promise.resolve(res.data.data);
-        }
-      }
-    });
+// store token scope
+const config = {
+    '/api/a' : 'BASE',
+    '/api/p' : 'PROJECT',
+    '/api/o' : 'PROJECT',
+    '/api/n' : 'PROJECT',
+    '/api/ui' : 'PROJECT',
+    '/api/t' : 'PROJECT'
 };
 
-exports.post = (route, payload) => {
-  let jwt = store.get('jwt');
-  let url = `${BASE_URL}${route}`;
-  let headers = {
-    'authorization': jwt
-  }
+function getTypeToken(route) {
+    for (let routeAlias in config) {
+        if (route.indexOf(routeAlias) !== -1) return config[routeAlias];
+    }
 
-  return axios.post(url, payload, {headers})
-    .then(res => {
-      if (!res.data) {
-        throw new Error('Something went wrong');
-      } else {
-        if (res.data.err) {
-          throw new Error(res.data.msg);
-        } else {
-          return Promise.resolve(res.data.data);
-        }
-      }
-    });
+    return 'BASE';
+}
+
+exports.get = (route, access_token = tokenGetter(getTypeToken(route))) => {
+    let url = `${BASE_URL}${route}`;
+    let headers = {
+        'authorization': access_token
+    };
+
+    return axios.get(url, {headers})
+        .then(handleResponse);
 };
 
-exports.put = (route, payload) => {
-  let jwt = store.get('jwt');
-  let url = `${BASE_URL}${route}`;
-  let headers = {
-    'authorization': jwt
-  }
+exports.post = (route, payload, access_token = tokenGetter(getTypeToken(route))) => {
+    let url = `${BASE_URL}${route}`;
+    let headers = {
+        'authorization': access_token
+    };
 
-  return axios.put(url, payload, {headers})
-    .then(res => {
-      if (!res.data) {
-        throw new Error('Something went wrong');
-      } else {
-        if (res.data.err) {
-          throw new Error(res.data.msg);
-        } else {
-          return Promise.resolve(res.data.data);
-        }
-      }
-    });
+    return axios.post(url, payload, {headers})
+        .then(handleResponse);
 };
 
-exports.delete = (route, access_token) => {
-  let jwt = store.get('jwt');
-  let url = `${BASE_URL}${route}`;
-  let headers = {
-    'authorization': access_token || jwt
-  }
+exports.put = (route, payload, access_token = tokenGetter(getTypeToken(route))) => {
+    let url = `${BASE_URL}${route}`;
+    let headers = {
+        'authorization': access_token
+    };
 
-  return axios.delete(url, {headers})
-    .then(res => {
-      if (!res.data) {
-        throw new Error('Something went wrong');
-      } else {
-        if (res.data.err) {
-          throw new Error(res.data.msg);
+    return axios.put(url, payload, {headers})
+        .then(handleResponse);
+};
+
+exports.delete = (route, access_token = tokenGetter(getTypeToken(route))) => {
+    let url = `${BASE_URL}${route}`;
+    let headers = {
+        'authorization': access_token || access_token
+    };
+
+    return axios.delete(url, {headers})
+        .then(handleResponse);
+};
+
+function handleResponse(res) {
+    if (!res.data) {
+        return Promise.reject(new Error('Something went wrong'));
+    } else {
+        if (res.data.success) {
+            return Promise.resolve(res.data.data);
         } else {
-          return Promise.resolve(res.data.data);
+            return Promise.reject(new Error(res.data.reason));
         }
-      }
-    });
+    };
 };
 
 exports.getWithoutAuth = (route) => {
-  let url = BASE_URL + route;
-  return axios.get(url);
-};
-
-exports.getGraphApi = (route, page_id) => {
-  let url = `https://graph.facebook.com/v2.10` + route + `&access_token=`;
-  let access_token = store.get(page_id);
-  if (!access_token) {
-    return DashboardAPI.getAccessToken(page_id)
-      .then((res) => {
-        url = url + res.access_token;
-        store.set(page_id, res.access_token);
-        return axios.get(url);
-      })
-  } else {
-    url += access_token
+    let url = BASE_URL + route;
     return axios.get(url);
-  }
 };
