@@ -1,91 +1,84 @@
-'use strict';
-
-let apiSender = require('ApiSender');
+let apiSender = require('./ApiSender');
 
 module.exports = {
-	getConversations: function (alias, next) {
-		let route = `/api/projects/${alias}/conversations`;
-		if (next) route += `?next=${next}`;
-		return apiSender.get(route);
-	},
-	getMessageInbox: function (inbox_id, next) {
-		let route = `/api/inboxes/${inbox_id}/messages`;
-		if (next) route += `?next=${next}`;
-		return apiSender.get(route);
-	},
-	getReplyComment: function (comment_id, next) {
-		let route = `/api/comments/${comment_id}/comments`;
-		if (next) route += `?next=${next}`;
-		return apiSender.get(route);
-	},
-	postSeenCmt: function (comment_id) {
-		let route = `/api/comments/${comment_id}/seen`;
-		return apiSender.post(route);
-	},
-	postSeenInbox: function (inbox_id) {
-		let route = `/api/inboxes/${inbox_id}/seen`;
-		return apiSender.post(route);
-	},
-	postRepInboxMsg: function (inbox_id, message) {
-		let route = `/api/inboxes/${inbox_id}/sendmsg`;
-		let payload = { message };
-		return apiSender.post(route, payload);
-	},
-	postRepCmtMsg: function (cmt_id, message, attachment_url) {
-		let route = `/api/comments/${cmt_id}/sendmsg`;
-		let payload = {};
-		if (message) {
-			payload.message = message;
-		}
-		if (attachment_url) {
-			payload.attachment_url = attachment_url;
-		}
-		return apiSender.post(route, payload);
-	},
-	getPostInfo: (post_id) => {
-		let route = `/api/posts/${post_id}`;
-		return apiSender.get(route);
-	},
-	getAccessToken: (page_id) => {
-		let route = `/api/pages/${page_id}/access-token`;
-		return apiSender.get(route);
-	},
-	getMessageAttachment: (msg_id, page_id) => {
-		let route = `/${msg_id}?fields=attachments`;
-		return apiSender.getGraphApi(route, page_id);
-	},
-	getCommentAttachment: (comment_id, page_id) => {
-		let route = `/${comment_id}?fields=attachment`;
-		return apiSender.getGraphApi(route, page_id);
-	},
-	getMessageShare: (msg_id, page_id) => {
-		let route = `/${msg_id}?fields=shares{link}`;
-		return apiSender.getGraphApi(route, page_id);
-	},
-	postPrivateReplyMessage: (comment_id, message) => {
-		let route = `/api/comments/${comment_id}/private_replies`;
-		let payload = { message };
-		return apiSender.post(route, payload);
-	},
-	getProjectTags: (alias) => {
-		let route = `/api/projects/${alias}/tags`;
-		return apiSender.get(route);
-	},
-	createTagConversation: (alias, conversation_id, tag_id) => {
-		let route = `/api/projects/${alias}/conversations/${conversation_id}/tags`;
-		let payload = {tag_id: tag_id};
-		return apiSender.post(route, payload);
-	},
-	deleteTagConversation: (alias, conversation_id, tag_id) => {
-		let route = `/api/projects/${alias}/conversations/${conversation_id}/tags/${tag_id}`;
-		return apiSender.delete(route);
-	},
-	likeMessage: (comment_id) => {
-		let route = `/api/comments/${comment_id}/like`;
-		return apiSender.post(route);
-	},
-	unlikeMessage: (comment_id) => {
-		let route = `/api/comments/${comment_id}/unlike`;
-		return apiSender.post(route);
-	}
-}
+    getConversations: function (alias, next, query) {
+        let route = `/api/p/conversations`;
+        if (next) route += `?next=${next}&limit=30`;
+        else route += '?limit=30';
+        if (query) {
+            for (let prop in query) {
+                if (query.hasOwnProperty(prop)) route += `&${prop}=${query[prop]}`;
+            }
+        }
+        return apiSender.get(route);
+    },
+    checkUnreadComment: function () {
+        let route = `/api/p/comments/has-unread`;
+        return apiSender.get(route);
+    },
+    checkUnreadInbox: function () {
+        let route = `/api/p/inboxes/has-unread`;
+        return apiSender.get(route);
+    },
+    getMessages: function (type, msg_id, next) {
+        let route;
+        if (type === "comment") route = `/api/p/comments/${msg_id}/reply-comments`;
+        else route = `/api/p/inboxes/${msg_id}/messages`;
+        if (next) route += `?next=${next}`;
+        return apiSender.get(route);
+    },
+    postSeenCmt: function (comment_id) {
+        let route = `/api/p/comments/${comment_id}`;
+        let payload = {seen: true};
+        return apiSender.put(route, payload);
+    },
+    postSeenInbox: function (inbox_id) {
+        let route = `/api/p/inboxes/${inbox_id}`;
+        let payload = {seen: true};
+        return apiSender.put(route, payload);
+    },
+    postRepInboxMsg: function (inbox_id, message) {
+        let route = `/api/p/inboxes/${inbox_id}/messages`;
+        let payload = {message};
+        return apiSender.post(route, payload);
+    },
+    postRepCmtMsg: function (cmt_id, message, attachment_url) {
+        let route = `/api/p/comments/${cmt_id}/reply-comments`;
+        let payload = {};
+        if (message) payload.message = message;
+        if (attachment_url) payload.attachment_url = attachment_url;
+        return apiSender.post(route, payload);
+    },
+    blockCustomer: (page_id, customer_fb_id, state) => {
+        let route = `/api/p/pages/${page_id}/customers/${customer_fb_id}`;
+        let payload = {
+            block: state
+        };
+        return apiSender.put(route, payload);
+    },
+    postPrivateReplyMessage: (comment_id, message) => {
+        let route = `/api/p/comments/${comment_id}/private-replies`;
+        let payload = {message};
+        return apiSender.post(route, payload);
+    },
+    likeMessage: (comment_id, state) => {
+        let route = `/api/p/comments/${comment_id}`;
+        let payload = {like: state};
+        return apiSender.put(route, payload);
+    },
+    updateExpiredAttachmentMsg: (type, msg_id, conv_id) => {
+        let route, payload;
+        if (type === "comment") {
+            route = `/api/p/comments/${msg_id}`;
+            payload = {attachment: true};
+        } else {
+            route = `/api/p/inboxes/${conv_id}/messages/${msg_id}`;
+            payload = {attachments: true};
+        }
+        return apiSender.put(route, payload);
+    },
+    getConversation: (conv_id) => {
+        let route = `/api/p/conversations/${conv_id}`;
+        return apiSender.get(route);
+    }
+};
