@@ -1,14 +1,12 @@
-'use strict';
-
-let APISender = require('ApiSender');
+let APISender = require('./ApiSender');
 
 module.exports = {
-	getPages: function () {
-		let route = `/api/pages`;
-		return APISender.get(route);
-	},
-	activePage: function (pageid) {
-		let route = `/api/pages/${pageid}/active`;
-		return APISender.post(route);
-	}
+  getPages: function () {
+    let route = `/api/a/pages`;
+    return APISender.get(route);
+  },
+  activePage: function (pageid) {
+    let route = `/api/pages/${pageid}/active`;
+    return APISender.post(route);
+  }
 };
