@@ -1,72 +1,40 @@
-'use strict';
-
-let apiSender = require('ApiSender');
-import uuid from 'uuid';
+let apiSender = require('./ApiSender');
 
 module.exports = {
-  getPostsOfPage: (page_id, next) => {
-    let route = `/api/pages/${page_id}/posts`;
-    if (next) route += `?next=${next}`;
+    getPostsOfPage: (page_id, next) => {
+        let route = `/api/p/pages/${page_id}/posts`;
+        if (next) route += `?next=${next}`;
 
-    return apiSender.get(route);
-  },
-  getPostsOfProject: (project_alias, next) => {
-    let route = `/api/projects/${project_alias}/posts`;
-    if (next) route += `?next=${next}`;
+        return apiSender.get(route);
+    },
+    getPostsOfProject: (next) => {
+        let route = `/api/p/posts`;
+        if (next) route += `?next=${next}`;
 
-    return apiSender.get(route);
-  },
-  hideComment: (post_id, hide) => {
-    let route;
-    if (hide) route = `/api/posts/${post_id}/hidecomment`;
-    else route = `/api/posts/${post_id}/unhidecomment`;
-
-    return apiSender.post(route, {});
-  },
-  getMockPostsOfProject: () => {
-    return Promise.resolve(
-      [
-        {
-          "created_time": "2017-09-25T16:43:21+0000",
-          "message": "adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô",
-          fb_id: uuid()
-        },
-        {
-          "created_time": "2017-09-25T16:43:21+0000",
-          "message": "adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô",
-          fb_id: uuid()
-        },
-        {
-          "created_time": "2017-09-25T16:43:21+0000",
-          "message": "adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô",
-          fb_id: uuid()
-        },
-        {
-          "created_time": "2017-09-25T16:43:21+0000",
-          "message": "adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô",
-          fb_id: uuid()
-        },
-        {
-          "created_time": "2017-09-25T16:43:21+0000",
-          "message": "adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô",
-          fb_id: uuid()
-        },
-        {
-          "created_time": "2017-09-25T16:43:21+0000",
-          "message": "adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô",
-          fb_id: uuid()
-        },
-        {
-          "created_time": "2017-09-25T16:43:21+0000",
-          "message": "adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô",
-          fb_id: uuid()
-        },
-        {
-          "created_time": "2017-09-25T16:43:21+0000",
-          "message": "adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô, adTenSon ông ơi ra đây mà xem được giải oan rồi này. tôi biết ông vô tội mà :))) #Tô",
-          fb_id: uuid()
-        }
-      ]
-    )
-  }
+        return apiSender.get(route);
+    },
+    getPostInfo: (page_fb_id, post_id) => {
+        let route = `/api/p/pages/${page_fb_id}/posts/${post_id}`;
+        return apiSender.get(route);
+    },
+    hideComment: (post_id, hide) => {
+        let route = `/api/p/posts/${post_id}`;
+        let payload = {hide_comment: hide};
+        return apiSender.put(route, payload);
+    },
+    hidePhoneComment: (post_id, hide) => {
+        let route = `/api/p/posts/${post_id}`;
+        let payload = {hide_phone: hide};
+        return apiSender.put(route, payload);
+    },
+    updateExpiredAttachmentPost: (post_id) => {
+        let route = `/api/p/posts/${post_id}`;
+        let payload = {attachments: true};
+        return apiSender.put(route, payload);
+    },
+    addNewPost: (fb_ids, content) => {
+        let route = `/api/p/pages/posts`;
+        let payload = {page_fb_ids: fb_ids, content: content};
+        return apiSender.post(route, payload);
+    }
 };
