@@ -1,8 +1,6 @@
-'use strict';
-
-import apiSender from 'ApiSender';
+import apiSender from './ApiSender';
 
 exports.verifyAccessToken = (access_token) => {
-  const route = `/api/tokens`;
+  const route = `/api/a/tokens`;
   return apiSender.get(route, access_token);
-}
+};
